requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.2
APScheduler==3.11.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.10.0
mplcursors==0.6
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
twilio==9.4.1
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.18.3
